Quantized System Solver

Taylor0

Problem

Algorithms

Algorithms:

qss1(),qss2()

QuantizedSystemSolver.solve — Method

solve(prob::NLODEProblem{PRTYPE,T,Z,D,CS},tspan::Tuple{Float64, Float64};sparsity::Val{Sparsity}=Val(false)::Float64,saveat=1e-9::Float64::Float64,abstol=1e-4::Float64,reltol=1e-3::Float64,maxErr=Inf::Float64) where {PRTYPE,T,Z,D,CS,Sparsity}

dispatches on a specific integrator based on the algorithm provided. After the simulation, the solution is returned as a Solution object.

back to top Quantized System Solver