Available Algorithms

QuantizedSystemSolver.qss1 — Method

qss1() calls the explicit quantized state system solver with order 1

QuantizedSystemSolver.qss2 — Method

qss2() calls the explicit quantized state system solver with order 2

QuantizedSystemSolver.liqss1 — Method

liqss1() calls the imlicit quantized state system solver with order 1.

QuantizedSystemSolver.liqss2 — Method

liqss2() calls the imlicit quantized state system solver with order 2.

QuantizedSystemSolver.nmliqss1 — Method

nmliqss1() calls the modified imlicit quantized state system solver with order 1. It is efficient when the system contains large entries outside the main diagonal of the Jacobian .

QuantizedSystemSolver.nmliqss2 — Method

nmliqss2() calls the modified imlicit quantized state system solver with order 2. It is efficient when the system contains large entries outside the main diagonal of the Jacobian .