Problem

Problem extension can be achieved easily via PRTYPE which is of type Val, or another subtype of the superclass can be created.

QuantizedSystemSolver.NLODEProblem — Type

NLODEProblem{PRTYPE,T,Z,Y,CS} This is superclass for all NLODE problems. It is parametric on these:

- The problem type PRTYPE.

- The number of continuous variables T

- The number of events (zero crossing functions) Z

- The actual number of events (an if-else statment has one zero crossing functions and two events) Y

- The cache size CS.

QuantizedSystemSolver.NLODEContProblem — Type

NLODEContProblem{PRTYPE,T,Z,Y,CS} A struct that holds the Problem of a system of ODEs. It has the following fields:

- prname: The name of the problem

- prtype: The type of the problem

- a: The size of the problem

- b: The number of zero crossing functions

- c: The number of discrete events

- cacheSize: The size of the cache

- initConditions: The initial conditions of the problem

- eqs: The function that holds all the ODEs

- jac: The Jacobian dependency

- SD: The state derivative dependency

- exactJac: The exact Jacobian function

- jacDim: The Jacobian dimension (if sparsity to be exploited)